MapCreator/requirement.txt
bidict==0.22.1
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-SocketIO==5.3.3
folium==0.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
numpy==1.24.2
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
